flow-based codezoo/pMF-main/pyproject.toml
[project]
name = "pmf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "datasets>=4.5.0",
    "lpips>=0.1.4",
    "numpy>=2.4.2",
    "pyyaml>=6.0.3",
    "tensorboard>=2.20.0",
    "timm>=1.0.24",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "tqdm>=4.67.2",
    "transformers>=5.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
